Rewriting core library to allow type definitions to also act as packing/unpacking objects
