+++ packtype/struct.py
@@ -1,4 +1,4 @@
-# Copyright 2021, Peter Birch, mailto:peter@lightlogic.co.uk
+# Copyright 2023, Peter Birch, mailto:peter@intuity.io
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,74 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .container import Container
-from .enum import Enum
-from .offset import Offset
-from .scalar import Scalar
-from .typedef import Typedef
+from .assembly import PackedAssembly
 
-class Struct(Container):
-    """ Packed data structure formed of scalars, enumerations, and other types """
-
-    FROM_LSB = "FROM_LSB"
-    FROM_MSB = "FROM_MSB"
-
-    def __init__(self, name, fields, desc=None, width=None, pack=FROM_LSB):
-        """ Initialise structure with name and fields
-
-        Args:
-            name  : Name of the container
-            fields: Dictionary of fields
-            desc  : Optional description
-            width : Bit width (if omitted is calculated from sum of fields)
-        """
-        # Perform container construction
-        from .union import Union
-        super().__init__(name, fields, desc=desc, width=width, legal=[
-            Enum, Struct, Union, Scalar, Typedef
-        ])
-        # Sanity check width
-        assert (width is None) or (isinstance(width, int) or width >= 0), \
-            f"Width must be None or a positive integer, not '{width}'"
-        # Check the packing mode
-        assert pack in (Struct.FROM_LSB, Struct.FROM_MSB), \
-            f"Unknown packing mode '{pack}' for {name}"
-        assert (pack == Struct.FROM_LSB) or (width is not None), \
-            "When packing downwards from the MSB, the width must be specified"
-        self._pt_pack = pack
-        # Determine the position of each field
-        if pack == Struct.FROM_MSB:
-            next_msb = width - 1
-            for field in sorted(self._pt_values(), key=lambda x: x._pt_id):
-                base_lsb = next_msb - field._pt_width + 1
-                if field._pt_lsb is None:
-                    field._pt_lsb = base_lsb
-                elif isinstance(field._pt_lsb, Offset):
-                    field._pt_lsb = base_lsb - field._pt_lsb.value
-                else:
-                    assert field._pt_lsb <= base_lsb, \
-                        f"Field '{field._pt_name}' of {self._pt_name} specifies " \
-                        f"an out-of-order LSB ({field._pt_lsb}), was expecting a " \
-                        f"value less than or equal to {base_lsb}"
-                next_msb = (field._pt_lsb - 1)
-        else:
-            next_lsb = 0
-            for field in sorted(self._pt_values(), key=lambda x: x._pt_id):
-                if field._pt_lsb is None:
-                    field._pt_lsb = next_lsb
-                elif isinstance(field._pt_lsb, Offset):
-                    field._pt_lsb = next_lsb + field._pt_lsb.value
-                else:
-                    assert field._pt_lsb >= next_lsb, \
-                        f"Field '{field._pt_name}' of {self._pt_name} specifies " \
-                        f"an out-of-order LSB ({field._pt_lsb}), was expecting a " \
-                        f"value greater than or equal to {next_lsb}"
-                next_lsb = (field._pt_lsb + field._pt_width)
-        # Check the field width
-        fwidth = max((x._pt_width + x._pt_lsb) for x in self._pt_values())
-        assert width is None or fwidth <= width, (
-            f"Total field width {fwidth} is greater than the specified width {width}"
-        )
-        # Calculate the width of the struct
-        if not self._pt_width and len(self._pt_keys()) > 0:
-            self._pt_width = fwidth
+class Struct(PackedAssembly):
+    pass
