+++ packtype_old/templates/lang_cpp.cpp.mako
