+++ packtype_old/templates/lang_cpp.hpp.mako
